requirements.txt
google-generativeai>=0.8.0  # Direct Gemini API (simpler, fewer conflicts)
protobuf>=4.21,<5  # Pin to 4.x for compatibility
python-telegram-bot>=20.0
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
python-dotenv>=1.0.0

